Sync: 875. Koko Eating Bananas (Medium) - Leet2Git

+++ 875_Koko_Eating_Bananas.cpp
@@ -5,7 +5,7 @@ public:
 
         for(int pile : piles) {
             if(sum > h) break;
-            sum += ceil(pile / mid);
+            sum += ceil((double)pile / mid);
         }
 
         if(sum <= h) return true;
